LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/cssom-view/cssom-getBoundingClientRect-003-expected.txt

XXX

FAIL {Element,Range}.prototype.getBoundingClientRect on SVG <tspan>, Element assert_equals: left expected 108 but got 8
PASS {Element,Range}.prototype.getBoundingClientRect on SVG <tspan>, Range
